{{cookiecutter.project_slug}}/pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "FastAPI Template"
version = "0.1.0"
description = "FastAPI template with things you need."
authors = ["Paulo Sairel Don <hello@paulodon.com>"]
packages = [
    { include = "{{cookiecutter.project_slug}}", from = "." },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.89.1"}
fastapi-camelcase = "1.0.5"
typer = "0.7.0"
pandas = "1.5.3"
sentry-sdk = "1.14.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
loguru = "0.6.0"
sqlmodel = "0.0.8"
SQLAlchemy = "1.4.35"  # related issue: https://github.com/tiangolo/sqlmodel/issues/315
mysqlclient = "2.1.1"
fastapi-cache2 = {extras = ["redis"], version = "0.2.0"}
boto3 = "1.26.57"
toml = "0.10.2"
sshtunnel = "0.4.0"
toolz = "0.12.0"
mixer = "^7.2.2"
psycopg2 = "^2.9.5"
fastapi-auth0 = "^0.3.2"


[tool.poetry.dev-dependencies]
black = "22.12.0"
isort = "5.11.4"
flake8 = "6.0.0"
coverage = "7.1.0"
pre-commit = "3.0.0"
ipython = "8.8.0"
pytest-asyncio = "0.20.3"
pytest-recording = "0.12.1"

[tool.black]
line-length = 130
exclude = """
/(
    .git
  | .hg
  | .mypy_cache
  | .tox
  | .venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | .egg-info
  | core
  | data
  | alembic
)/
"""

[tool.isort]
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
default_section = "FIRSTPARTY"
line_length = 80
skip = "migrations,build,*.egg-info,core,venv,.tox,alembic"

[tool.coverage.run]
source = ["app"]
branch = true
omit = [""]

[tool.coverage.report]
exclude_lines = [
"# noqa",
"raise NotImplementedError"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry.scripts]
server = "{{cookiecutter.project_slug}}.main:dev_server"
test = "{{cookiecutter.project_slug}}.main:run_tests"
db = "{{cookiecutter.project_slug}}.main:init_db"
